move the type converter into a separate file.
